Tutor-de-equivalencia-logica/requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
requests==2.32.4
sympy==1.14.0
google-generativeai==0.8.5
spacy==3.8.7
lark==1.2.2
openai==1.30.5
python-dotenv==1.0.1
matplotlib==3.8.0